EntityLogAttr.getDisplayValue throws StringIndexOutOfBoundsException. #PL-4689

+++ modules/core/test/com/haulmont/cuba/security/EntityLogTest.java
@@ -6,9 +6,11 @@ package com.haulmont.cuba.security;
 
 import com.haulmont.cuba.core.*;
 import com.haulmont.cuba.core.global.AppBeans;
+import com.haulmont.cuba.core.global.Messages;
 import com.haulmont.cuba.security.app.EntityLogAPI;
 import com.haulmont.cuba.security.entity.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -18,10 +20,11 @@ import java.util.UUID;
  */
 public class EntityLogTest extends CubaTestCase {
 
-    private UUID leId;
-    private UUID laId;
+    private List<UUID> leId = new ArrayList<>();
+    private List<UUID> laId = new ArrayList<>();
 
     private UUID userId;
+    private UUID roleId;
 
     protected void setUp() throws Exception {
         super.setUp();
@@ -36,15 +39,27 @@ public class EntityLogTest extends CubaTestCase {
             q.executeUpdate();
 
             LoggedEntity le = new LoggedEntity();
-            leId = le.getId();
+            leId.add(le.getId());
             le.setName("sec$User");
             le.setAuto(true);
             em.persist(le);
 
             LoggedAttribute la = new LoggedAttribute();
-            laId = la.getId();
+            laId.add(la.getId());
             la.setEntity(le);
             la.setName("email");
+            em.persist(la);
+
+            le = new LoggedEntity();
+            leId.add(le.getId());
+            le.setName("sec$Role");
+            le.setAuto(true);
+            em.persist(le);
+
+            la = new LoggedAttribute();
+            laId.add(la.getId());
+            la.setEntity(le);
+            la.setName("type");
             em.persist(la);
 
             tx.commit();
@@ -143,22 +158,76 @@ public class EntityLogTest extends CubaTestCase {
         assertEquals(1, items.get(0).getAttributes().size());
     }
 
-    protected void tearDown() throws Exception {
+    public void testEnumDisplayValue() throws Exception {
         Transaction tx = persistence.createTransaction();
         try {
             EntityManager em = persistence.getEntityManager();
 
-            Query q = em.createNativeQuery("delete from SEC_LOGGED_ATTR where ID = ?");
-            q.setParameter(1, laId.toString());
-            q.executeUpdate();
+            Role role = new Role();
+            roleId = role.getId();
+            role.setName("role1");
+            role.setType(RoleType.READONLY);
+            em.persist(role);
 
-            q = em.createNativeQuery("delete from SEC_LOGGED_ENTITY where ID = ?");
-            q.setParameter(1, leId.toString());
-            q.executeUpdate();
-            
-            q = em.createNativeQuery("delete from SEC_USER where ID = ?");
-            q.setParameter(1, userId.toString());
-            q.executeUpdate();
+            tx.commit();
+        } finally {
+            tx.end();
+        }
+
+        List<EntityLogItem> items;
+        tx = persistence.createTransaction();
+        try {
+            EntityManager em = persistence.getEntityManager();
+            TypedQuery<EntityLogItem> query = em.createQuery(
+                    "select i from sec$EntityLog i where i.entity = ?1 and i.entityId = ?2", EntityLogItem.class);
+            query.setParameter(1, "sec$Role");
+            query.setParameter(2, roleId);
+            items = query.getResultList();
+
+            tx.commit();
+        } finally {
+            tx.end();
+        }
+        assertNotNull(items);
+        assertEquals(1, items.size());
+        assertNotNull(items.get(0).getAttributes());
+        assertEquals(1, items.get(0).getAttributes().size());
+
+        EntityLogAttr attr = items.get(0).getAttributes().iterator().next();
+
+        Messages messages = AppBeans.get(Messages.NAME);
+        assertEquals(messages.getMessage(RoleType.READONLY), attr.getDisplayValue());
+    }
+
+    protected void tearDown() throws Exception {
+        Transaction tx = persistence.createTransaction();
+        try {
+            EntityManager em = persistence.getEntityManager();
+            Query q;
+
+            for (UUID uuid : laId) {
+                q = em.createNativeQuery("delete from SEC_LOGGED_ATTR where ID = ?");
+                q.setParameter(1, uuid.toString());
+                q.executeUpdate();
+            }
+
+            for (UUID uuid : leId) {
+                q = em.createNativeQuery("delete from SEC_LOGGED_ENTITY where ID = ?");
+                q.setParameter(1, uuid.toString());
+                q.executeUpdate();
+            }
+
+            if (userId != null) {
+                q = em.createNativeQuery("delete from SEC_USER where ID = ?");
+                q.setParameter(1, userId.toString());
+                q.executeUpdate();
+            }
+
+            if (roleId != null) {
+                q = em.createNativeQuery("delete from SEC_ROLE where ID = ?");
+                q.setParameter(1, roleId.toString());
+                q.executeUpdate();
+            }
 
             tx.commit();
         } finally {
